show margin in elev plot

+++ src/mwp-plot-elevations/gnuplot.go
@@ -27,7 +27,9 @@ func Gnuplot_mission(mpts []Point, gnd []int) {
 	minz := 99999
 
 	tmpdir, err := ioutil.TempDir("", ".mplot")
-	defer os.RemoveAll(tmpdir)
+	if os.Getenv("KEEP_TEMPS") == "" {
+		defer os.RemoveAll(tmpdir)
+	}
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -103,7 +105,7 @@ set yrange [ `)
 	}
 	fmt.Fprintf(w, "plot \"%s\" using 1:2 t \"Terrain\" w filledcurve y1=%d lt -1 lw 2  lc rgb \"web-green\", \"%s\" using 1:2 t \"Mission\" w lines lt -1 lw 2  lc rgb \"red\"", tfname, minz, mfname)
 	if Conf.Margin != 0 {
-		fmt.Fprintf(w, ", \"%s\" using 1:3 t \"Margin\" w lines lt -1 lw 2  lc rgb \"web-blue\"", tfname)
+		fmt.Fprintf(w, ", \"%s\" using 1:3 t \"Margin %dm\" w lines lt -1 lw 2  lc rgb \"web-blue\"", tfname, Conf.Margin)
 	}
 	if Conf.Output != "" {
 		fmt.Fprintf(w, ", \"%s\" using 1:3 t \"Adjust\" w lines lt -1 lw 2  lc rgb \"orange\"", mfname)
